generalize init plugins
