+++ app/kumactl/cmd/install/testdata/install-control-plane.dump-values.yaml
@@ -721,6 +721,7 @@ postgres:
 
 # @ignored for helm-docs
 plugins:
+  resources: {}
   policies:
     meshaccesslogs: true
     meshcircuitbreakers: true
