+++ pkg/core/bootstrap/bootstrap.go
@@ -75,8 +75,8 @@ func buildRuntime(appCtx context.Context, cfg kuma_cp.Config) (core_runtime.Runt
 	if err != nil {
 		return nil, err
 	}
-	core_apis.InitAllPolicies()
-	policies.InitPolicies(cfg.Policies.PluginPoliciesEnabled)
+	core_plugins.InitAll(core_apis.NameToModule)
+	core_plugins.Init(cfg.Policies.PluginPoliciesEnabled, policies.NameToModule)
 	builder.WithMultitenancy(multitenant.SingleTenant)
 	builder.WithPgxConfigCustomizationFn(config.NoopPgxConfigCustomizationFn)
 	for _, plugin := range core_plugins.Plugins().BootstrapPlugins() {
