--- pkg/core/resources/apis/imports.go
@@ -1,32 +1,0 @@
-// Generated by tools/policy-gen
-// Run "make generate" to update this file.
-
-package policies
-
-type PluginInitializer struct {
-	InitFn      func()
-	Initialized bool
-}
-
-var NameToModule = map[string]*PluginInitializer{}
-
-func InitAllPolicies() {
-	for _, initializer := range NameToModule {
-		if !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		}
-	}
-}
-
-func InitPolicies(enabledPluginPolicies []string) {
-	for _, policy := range enabledPluginPolicies {
-		initializer, ok := NameToModule[policy]
-		if ok && !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		} else {
-			panic("policy " + policy + " not found")
-		}
-	}
-}
