+++ pkg/core/resources/apis/modules.go
@@ -1,0 +1,9 @@
+// Generated by tools/policy-gen
+// Run "make generate" to update this file.
+package policies
+
+import (
+	"github.com/kumahq/kuma/pkg/core/plugins"
+)
+
+var NameToModule = map[string]*plugins.PluginInitializer{}
