+++ pkg/plugins/policies/modules.go
@@ -1,9 +1,9 @@
 // Generated by tools/policy-gen
 // Run "make generate" to update this file.
-
 package policies
 
 import (
+	"github.com/kumahq/kuma/pkg/core/plugins"
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshaccesslog"
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshcircuitbreaker"
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshfaultinjection"
@@ -20,12 +20,7 @@ import (
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshtrafficpermission"
 )
 
-type PluginInitializer struct {
-	InitFn      func()
-	Initialized bool
-}
-
-var NameToModule = map[string]*PluginInitializer{
+var NameToModule = map[string]*plugins.PluginInitializer{
 	"meshaccesslogs":              {InitFn: meshaccesslog.InitPlugin, Initialized: false},
 	"meshcircuitbreakers":         {InitFn: meshcircuitbreaker.InitPlugin, Initialized: false},
 	"meshfaultinjections":         {InitFn: meshfaultinjection.InitPlugin, Initialized: false},
@@ -40,25 +35,4 @@ var NameToModule = map[string]*PluginInitializer{
 	"meshtimeouts":                {InitFn: meshtimeout.InitPlugin, Initialized: false},
 	"meshtraces":                  {InitFn: meshtrace.InitPlugin, Initialized: false},
 	"meshtrafficpermissions":      {InitFn: meshtrafficpermission.InitPlugin, Initialized: false},
-}
-
-func InitAllPolicies() {
-	for _, initializer := range NameToModule {
-		if !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		}
-	}
-}
-
-func InitPolicies(enabledPluginPolicies []string) {
-	for _, policy := range enabledPluginPolicies {
-		initializer, ok := NameToModule[policy]
-		if ok && !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		} else {
-			panic("policy " + policy + " not found")
-		}
-	}
 }
