+++ tools/policy-gen/generate-policy-import.sh
@@ -3,7 +3,7 @@
 set -e
 GO_MODULE=$1
 POLICIES_DIR=$2 # pkg/plugins/policies or pkg/core/resources/apis
-IMPORTS_FILE="${POLICIES_DIR}/imports.go"
+IMPORTS_FILE="${POLICIES_DIR}/modules.go"
 
 imports=$(
   for i in "${@:3}"; do
@@ -25,41 +25,15 @@ done)
 
 echo "// Generated by tools/policy-gen
 // Run \"make generate\" to update this file.
-
 package policies
 
 import (
-  $imports
+	\"github.com/kumahq/kuma/pkg/core/plugins\"
+	$imports
 )
 
-type PluginInitializer struct {
-	InitFn      func()
-	Initialized bool
-}
-
-var NameToModule = map[string]*PluginInitializer{
+var NameToModule = map[string]*plugins.PluginInitializer{
   $mappings
-}
-
-func InitAllPolicies() {
-	for _, initializer := range NameToModule {
-		if !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		}
-	}
-}
-
-func InitPolicies(enabledPluginPolicies []string) {
-	for _, policy := range enabledPluginPolicies {
-		initializer, ok := NameToModule[policy]
-		if ok && !initializer.Initialized {
-			initializer.InitFn()
-			initializer.Initialized = true
-		} else {
-			panic(\"policy \" + policy + \" not found\")
-		}
-	}
 }
 " > "${IMPORTS_FILE}"
